Insert space after period on landing page
There is no space between the "restrict you." and "Upload" in the rendered html.  It seems that if there is not an explicit space (not just a newline) between the period and the `<strong>` tag, the space is deleted in the rendered html.  Reformatting the paragraph ensures that the newline falls in normal text, which seems to not have the problem (evidenced by other paragraphs).