wip poc: xterm: hotkey for editor change
Github issue: #618 (also #3897)

The purpose is to have a global hot key for switching between open files of a project.

This patch provides it only for xterminals.

I tried installing a keydown event handler on a top level div but it missed all the events
that the xterm was grabbing, so I went with changing the event handler for the xterm.

For other editors this might have to be done as well - this is an unfortunate solution
since it is invasive, a single top level global key handler would be easier to reason about.

This functionality can however be extracted and reused between editors.

I'd like feedback:
 - is this approach reasonable?
 - do you have a different approach in mind for the issue?