hub server -- never let static/* fails fall back to nextjs server
- I think when doing development a single false request (maybe due
to caching) to static/* would fall back to nextjs, which would
then take 10s to generate the 404 page for the first time, thus
breaking everything. Instead, we just immediately generate a 404
page in such cases.

This should only impact cc-in-cc dev, though it might help a tiny bit
with cocalc-docker as well.