jupyter service: better kernel pool management and cell limits
This commit addresses an issue with the management of kernel pools, where a
pool could stick around indefinitely without any means to clear it. We added
a timeout to each pool (soon to be configurable via admin settings) to clear
unused kernels.

Additionally, this commit adds better configurable support for cell execution
limits, allowing for a maximum time limit and maximum output size limit per
cell. These limits are wired up to be settable on a per-project level (with different
default limits for the general public project), but for now that will
not be user facing or configurable.  Basically, we're laying the
foundations so we don't have to update the project image later.