do not autoshow run button output for jupyter
Important because often the cached output that doesn't work instead of the correct actual output in the file itself. Thus: for jupyter never try to get and show the cached output proactively. ONLY do that for markdown. It's an easy change.    We have to do this, e.g,. because of network or raw input etc.