Add "expand-arrows" icon, reformat grid row layout, and add string tagging to "accounts" schema
This commit adds the "ArrowsAltOutlined" icon to the `icon.tsx` file and applies it to the expand button on the "Grid" view in the `grid.tsx` file. It also reformats the layout of grid rows, adding a button to expand the row details and moving the index column further to the left of the row. Lastly, the commit adds string tagging to the `accounts` schema, with a custom rendering type. Note that it's not editable yet really, despite saying it is...