fix #7144 -- Subscription renewal is seriously broken
- subscription renewal happened fine, but a few days later the
attempt to collect the automatic payment would be skipped because
the amount due was below the min allowed payment amount of $5. E.g.
if there was a $3.07/month subscription a pending payment got created
and the subscription was renewed, but no payment gets made.
- then another maintenance task would cancel the subscription since
it wasn't paid.

This was caused by tension between collecting payments
for users with very small subscriptions, balanced with the substantial
transaction fees.  I've changed the code so that when anything at all
is due, if it is very small, instead the min allowed amount is
collected.  In the case of a $3.07/month subscription, an additional
$1.93 positive credit would result. That automatically gets applied
toward future subscription payments.   I could have instead tried
to allow a negative balance for a while, but that MASSIVELY complicates
a lot of other code, and would surely introduce many bugs.  This
solution was very simple to implement and test, so I'm going with it.