re-vamp terminals so that control messages are processed from only the leader client
why? I got curious if codex could once and for all solve the subtle problem of multiple collaborative users of a terminal at once causing control code issues. It's just really tedious and I have put it off forever, but the problem is if multiple people have the same terminal open and a process in the project sends a message about terminal capabilities, all user terminals answer, whereas only one should. The right solution involves delegating one client to be a temporary leader, doing health checks, etc. It's tedious. Also you have to know whether data from the terminal is from the user or in response to capability/control messages... and I have never known how to do that -- codex has read the source code (I guess) of xterm.js, and it knew an undocumented non-public API that provides that info. With some back and forth, it works. 🤷